% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{calculate_similarity_matrix}
\alias{calculate_similarity_matrix}
\title{Calculate Cosine Similarity Matrix}
\usage{
calculate_similarity_matrix(M)
}
\arguments{
\item{M}{Reference matrix with fluorophores in rows and detectors in columns.}
}
\value{
A square matrix of pairwise cosine similarities.
}
\description{
Calculate Cosine Similarity Matrix
}
