% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openpanel-package.R
\docType{package}
\name{openpanel-package}
\alias{openpanel}
\alias{openpanel-package}
\title{openpanel}
\description{
Interactive spectral flow cytometry panel design.
}
\author{
\strong{Maintainer}: Paul Heisig \email{pkheisig@gmail.com} (\href{https://orcid.org/0000-0002-8529-7944}{ORCID})

Authors:
\itemize{
  \item Paul Heisig \email{pkheisig@gmail.com} (\href{https://orcid.org/0000-0002-8529-7944}{ORCID})
}

}
\keyword{internal}
