% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{plot_similarity_matrix}
\alias{plot_similarity_matrix}
\title{Plot Cosine Similarity Matrix}
\usage{
plot_similarity_matrix(
  similarity_matrix,
  output_file = NULL,
  width = 180,
  height = 160
)
}
\arguments{
\item{similarity_matrix}{Matrix returned by \code{\link[=calculate_similarity_matrix]{calculate_similarity_matrix()}}.}

\item{output_file}{Optional path at which to save the plot.}

\item{width}{Plot width in millimetres.}

\item{height}{Plot height in millimetres.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Cosine Similarity Matrix
}
